backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Backend for CryptoShow"
requires-python = ">=3.12"
authors = [{ name = "Lukas Polak", email = "admin@lukaspolak.cz" }]

dependencies = [
    # FastAPI
    "fastapi[standard]",
    "requests",
    "prometheus-fastapi-instrumentator",
    "flower",

    # CryptoBench (Tiny), ESM-2
    "biopython==1.85",
    "transformers==4.48.1",
    "torch==2.6.0",

    # Clustering
    "scikit-learn",

    # Sequence + Structure functionality
    "biotite",

    # Trajectory generation
    "numpy",
    "MDAnalysis",
    "gemmi",
    "scipy",

    # Async tasks
    "celery",
    "redis"
]

